+++ interpreter/value_fix128.go
@@ -534,7 +534,7 @@ func (Fix128Value) IsStorable() bool {
 	return true
 }
 
-func (v Fix128Value) Storable(_ atree.SlabStorage, _ atree.Address, _ uint64) (atree.Storable, error) {
+func (v Fix128Value) Storable(_ atree.SlabStorage, _ atree.Address, _ uint32) (atree.Storable, error) {
 	return v, nil
 }
