+++ interpreter/value_int8.go
@@ -598,7 +598,7 @@ func (v Int8Value) ConformsToStaticType(
 	return true
 }
 
-func (v Int8Value) Storable(_ atree.SlabStorage, _ atree.Address, _ uint64) (atree.Storable, error) {
+func (v Int8Value) Storable(_ atree.SlabStorage, _ atree.Address, _ uint32) (atree.Storable, error) {
 	return v, nil
 }
