+++ interpreter/value_word128.go
@@ -562,7 +562,7 @@ func (Word128Value) IsStorable() bool {
 	return true
 }
 
-func (v Word128Value) Storable(_ atree.SlabStorage, _ atree.Address, _ uint64) (atree.Storable, error) {
+func (v Word128Value) Storable(_ atree.SlabStorage, _ atree.Address, _ uint32) (atree.Storable, error) {
 	return v, nil
 }
